Introduced timeouts to Infrareddy read/write operations.
Also added try..catch to actual Read/Write to capture these conditions
and return something meaningful to the application.

This update may improve the experience for installers, and can be
shipped by itself, but an additional firmware update is also likely
required in order to fully mitigate the IR record crash scenario.